Fix Namespace issues & add supervisor repo into IoC
