+++ src/KP.BackEnd/Core/Repositories/ISchoolClassRepository.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using KP.BackEnd.Core.Models;
 
@@ -8,5 +9,6 @@ namespace KP.BackEnd.Core.Repositories
     {
         Task<SchoolClass> Find(Guid userId, Guid id);
         Task Add(SchoolClass schoolClass);
+        Task<IEnumerable<SchoolClass>> GetAll(Guid userId);
     }
 }
