+++ src/KP.BackEnd/Startup.cs
@@ -86,6 +86,7 @@ namespace KP.BackEnd
             services.AddScoped<IUnitOfWork, UnitOfWork>();
             services.AddScoped<ICardRepository, CardRepository>();
             services.AddScoped<ICommentRepository, CommentRepository>();
+            services.AddScoped<ISchoolClassRepository, SchoolClassRepository>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
